use framework to reduce ipa size
