--- common/common-public.h
@@ -1,4 +1,0 @@
-#include <string>
-
-std::string getInputMethods();
-void setLocale(const char *locale);
