+++ common/module.modulemap
@@ -1,3 +1,3 @@
-module FcitxCommon {
-  header "common-public.h"
+module Fcitx {
+  header "runtime-api.h"
 }
