+++ common/runtime_main.cpp
@@ -1,9 +1,49 @@
-#include "common.h"
 #include "../fcitx5/src/lib/fcitx/addonmanager.h"
+#include "common.h"
 #include "nativestreambuf.h"
+#include "runtime-api.h"
+#include "util.h"
 #include <fcitx-utils/i18n.h>
+#include <fcitx/inputmethodmanager.h>
 #include <filesystem>
+#include <nlohmann/json.hpp>
 #include <thread>
+
+void startFcitx(const char *appBundlePath, const char *xdgDataDirs,
+                const char *appGroupPath) {
+    if (instance) {
+        return;
+    }
+    setupFcitx(appBundlePath, xdgDataDirs, appGroupPath, true);
+}
+
+void setInputMethods(const char *s) {
+    std::string json = s;
+    dispatcher->schedule([=] {
+        auto &imMgr = instance->inputMethodManager();
+        auto group = imMgr.currentGroup();
+        auto &imList = group.inputMethodList();
+        imList.clear();
+        auto j = nlohmann::json::parse(json);
+        for (const auto &im : j) {
+            imList.emplace_back(im.get<std::string>());
+        }
+        imMgr.setGroup(group);
+        imMgr.save();
+    });
+}
+
+std::string getAllInputMethods() {
+    return with_fcitx([] {
+        auto j = nlohmann::json::array();
+        auto &imMgr = instance->inputMethodManager();
+        imMgr.foreachEntries([&j](const fcitx::InputMethodEntry &entry) {
+            j.push_back(jsonDescribeIm(&entry));
+            return true;
+        });
+        return j.dump();
+    });
+}
 
 namespace fs = std::filesystem;
