+++ patches/rime.patch
@@ -1,3 +1,13 @@
+diff --git a/CMakeLists.txt b/CMakeLists.txt
+index f8201d7..a4400c8 100644
+--- a/CMakeLists.txt
++++ b/CMakeLists.txt
+@@ -1,4 +1,4 @@
+-cmake_minimum_required(VERSION 3.6.0)
++cmake_minimum_required(VERSION 3.13)
+ 
+ project(fcitx5-rime VERSION 5.1.14)
+ 
 diff --git a/src/rimeengine.cpp b/src/rimeengine.cpp
 index cdaaa3d..c91da5c 100644
 --- a/src/rimeengine.cpp
