--- src/module.modulemap
@@ -1,3 +1,0 @@
-module Fcitx {
-  header "fcitx.h"
-}
